refactor: ReadApi 테스트 병합
