+++ backend/baton/src/test/java/touch/baton/document/profile/runner/read/RunnerProfileReadApiTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import touch.baton.config.RestdocsConfig;
+import touch.baton.domain.member.Member;
 import touch.baton.domain.runner.Runner;
 import touch.baton.domain.runner.controller.RunnerProfileController;
 import touch.baton.domain.runner.service.RunnerService;
@@ -19,9 +20,13 @@ import touch.baton.fixture.domain.TechnicalTagFixture;
 import java.util.List;
 import java.util.Optional;
 
+import static javax.swing.text.html.parser.DTDConstants.NUMBER;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.notNull;
 import static org.mockito.BDDMockito.when;
+import static org.mockito.Mockito.spy;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
@@ -30,6 +35,8 @@ import static org.springframework.restdocs.payload.JsonFieldType.STRING;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static touch.baton.fixture.vo.TagNameFixture.tagName;
 
 @WebMvcTest(RunnerProfileController.class)
@@ -76,5 +83,36 @@ class RunnerProfileReadApiTest extends RestdocsConfig {
                         )
                 ))
                 .andDo(print());
+    }
+
+    @DisplayName("러너 프로필 상세 조회 API")
+    @Test
+    void readRunnerProfile() throws Exception {
+        // given
+        final Member ethan = MemberFixture.createEthan();
+        final TechnicalTag javaTag = TechnicalTagFixture.createJava();
+        final TechnicalTag reactTag = TechnicalTagFixture.createReact();
+        final Runner runner = RunnerFixture.createRunner(ethan, List.of(javaTag, reactTag));
+        final Runner spyRunner = spy(runner);
+
+        // when
+        when(spyRunner.getId()).thenReturn(1L);
+        when(runnerService.readRunnerById(anyLong())).thenReturn(spyRunner);
+
+        // then
+        mockMvc.perform(get("/api/v1/profile/runner/{runnerId}", 1L))
+                .andExpect(status().isOk())
+                .andExpect(content().contentType(APPLICATION_JSON))
+                .andDo(restDocs.document(
+                        responseFields(
+                                fieldWithPath("runnerId").type(NUMBER).description("러너 식별자값"),
+                                fieldWithPath("name").type(STRING).description("러너 이름"),
+                                fieldWithPath("imageUrl").type(STRING).description("사용자 이미지"),
+                                fieldWithPath("githubUrl").type(STRING).description("깃허브 프로필 url"),
+                                fieldWithPath("introduction").type(STRING).description("소개"),
+                                fieldWithPath("company").type(STRING).description("소속"),
+                                fieldWithPath("technicalTags").type(ARRAY).description("기술 스택")
+                        )
+                ));
     }
 }
